+++ NEWS.md
@@ -3,6 +3,8 @@ PENDING: Dataiter 0.37
 
 * `DataFrame`: Add placeholder attributes for columns so that tab
   completion of columns as attributes at a shell works
+* `dt.from_string`: New function
+* `dt.to_string`: New function
 
 2022-10-16: Dataiter 0.36
 =========================